another way to process test case properties
