+++ src/extension/IServiceMessageFactory.cs
@@ -10,9 +10,7 @@
         IEnumerable<ServiceMessage> SuiteStarted(EventId eventId, Event testEvent);
 
         IEnumerable<ServiceMessage> SuiteFinished(EventId eventId, Event testEvent);
-
-        IEnumerable<ServiceMessage> SuiteProperties(EventId eventId, XmlNode testSuiteNode);
-
+    
         IEnumerable<ServiceMessage> FlowStarted(string flowId, string parentFlowId);
 
         IEnumerable<ServiceMessage> FlowFinished(string flowId);
