+++ src/extension/ServiceMessageFactory.cs
@@ -33,62 +33,9 @@
 
         public IEnumerable<ServiceMessage> SuiteFinished(EventId eventId, Event testEvent)
         {
-            if (_teamCityInfo.MetadataEnabled)
-            {
-                foreach (var message in Attachments(eventId, testEvent.TestEvent))
-                {
-                    yield return message;
-                }
-                foreach (var message in TestProperties(eventId, testEvent.TestEvent))
-                {
-                    yield return message;
-                }
-            }
             yield return new ServiceMessage(ServiceMessage.Names.TestSuiteFinished,
                 new ServiceMessageAttr(ServiceMessageAttr.Names.Name, _suiteNameReplacer.Replace(testEvent.Name)),
                 new ServiceMessageAttr(ServiceMessageAttr.Names.FlowId, eventId.FlowId));
-        }
-
-        public IEnumerable<ServiceMessage> SuiteProperties(EventId eventId, XmlNode testSuiteNode)
-        {
-          var properties = testSuiteNode.SelectNodes("properties/property");
-          var testCases = testSuiteNode.SelectNodes("test-case");
-
-          if (properties != null && testCases != null)
-          {
-            var props = new NameValueCollection();
-            foreach (var property in properties)
-            {
-              var propertyElement = property as XmlNode;
-              if (propertyElement == null)
-              {
-                continue;
-              }
-
-              var propertyName = propertyElement.GetAttribute("name") ?? string.Empty;
-              var propertyValue = propertyElement.GetAttribute("value") ?? string.Empty;
-              props.Add(propertyName, propertyValue);
-            }
-
-            foreach (var testCase in testCases)
-            {
-              var testCaseElement = testCase as XmlNode;
-              var fullName = testCaseElement.GetAttribute("fullname");
-              var flowId = testCaseElement.GetAttribute("id") ?? eventId.FlowId;
-              foreach (var name in props.AllKeys)
-              {
-                var attrs = new List<ServiceMessageAttr>
-                {
-                  new ServiceMessageAttr(ServiceMessageAttr.Names.FlowId, flowId),
-                  new ServiceMessageAttr(ServiceMessageAttr.Names.TestName, fullName),
-                  new ServiceMessageAttr(ServiceMessageAttr.Names.Name, name),
-                  new ServiceMessageAttr(ServiceMessageAttr.Names.Value, props[name])
-                };
-
-                yield return new ServiceMessage(ServiceMessage.Names.TestMetadata, attrs);
-              }
-            }
-          }
         }
 
         public IEnumerable<ServiceMessage> FlowStarted(string flowId, string parentFlowId)
